refactor: Extract game notification logic to useGameNotifications hook
